tools/cli/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "raca"
version = "0.2.0"
description = "RACA — SSH lifecycle for research clusters"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "pexpect>=4.8",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
raca = "raca.cli:main"
